All source related features updated and the logic implemented
Except Initial soldier, limit for soldier per tile and buildings level

+++ UntilTheLastColorStand/core/src/main/java/io/github/some_example_name/resources/FoodResource.java
@@ -3,11 +3,11 @@ package io.github.some_example_name.resources;
 and methods for food related features */
 
 public class FoodResource extends Resource {
-    public final int RECRUIT;
-    public final int TILE;
-    public final int MAINTAIN;
+    public final double RECRUIT;
+    public final double TILE;
+    public final double MAINTAIN;
 
-    public FoodResource(int value,int RECRUIT, int TILE, int MAINTAIN) {
+    public FoodResource(int value,double RECRUIT, double TILE, double MAINTAIN) {
         super(value);
         this.RECRUIT = RECRUIT;
         this.TILE = TILE;
